(bug fix) fix errors from moving files around, dynamic rightIcon
